Add new "ignore_checkpoints_on_startup_path_globs" config option.
With this config option user can specify path globs for log file path
which checkpoint data should be ignored when reading checkpoints data on
agent start up.

This comes handy in situations where we only want to preserve some
checkpoints across restart.

An example of that is Kubernetes agent deployment where we persist
checkpoints across restart to avoid duplicated data for container logs
which are stored on the host and are persisted across restarts.

On the other hand, we also have all other log files which are ephemeral
to the agent container (anything in /var/log/scalyr-agent-2/*.log) for
which we don't want to persist checkpoints across restarts since those
checkpoints would refer to previous container logs which are not there
anymore.