Don't delete log files at the end of the tests to make troubleshooting failures easier.
If those tests fail, exceptions don't bubble it up and are only
contained inside the log files - this means that if we delete the log
files, it's almost impossible to figure out test case failure root
cause.

Sadly there is no easy way to detect test failure with pytest and bubble
that up to the tearDown() method so we need to go with this approach
where we always preserve those log files.