Allow custom timestamp to be specified for emitted metric log lines (#856)
* Update monitor related "emit_value()" method so it allows user to
specify a custom timestamp to use for that emitted metric value (by
default it will use a value of a time when Record object is created).

This comes handy when we want multiple metrics (e.g. all the metrics
which are part of the same scrape / batch) to contain the same
timestamp.

* Update OpenMetrics monitor so we use the same timestamp for all the
metrics which are parsed as part of the same batch in case a metric
itself doesn't include a timestamp.