Customize packages installed on windows in python 3.10
We need to override the version of the pywin32 package installed on
windows in python 3.10. To accomplish this, I'm trying an experiment
using "factor-conditional" settings feature in tox. The package
tox-gh-actions configures factors based on job configuration, and then
we use the python version and platform factors to cause tox to install
the package version we need, where we need it.

Trying to do this in a manner that doesn't break the circleci tests.

CT-574