CT-574 Migrate Unittests from CircleCI to GHA (#950)
* First pass at Unittest GHA workflow

CT-574

* Run unittests on multiple versions of python

CT-574

* Add python3.10 tox unit test target

CT-574

* Create test reports

CT-574

* Add missing permissions for test report

CT-574

* Add pytest plugin to annotate failures

CT-574

* Each unit test job should use a separate comment

This is to prevent each job from overwriting reports from other jobs.

CT-574

* Begin multi-platform unittest support by adding macos

CT-574

* Fix publish unit test result action for other platforms

The `publish-unit-test-result-action` action runs in a docker container,
and as a result an alternate version of the action must be run on macos
and windows.

Also, don't publish tests if the suite is canceled. There is no point.

CT-574

* Add windows platform to test matrix

Add windows platform to test matrix to see what fails.
Also fix name collision in upload of pytest artifacts.

CT-574

* Fix unittests on python 3.7

Unittests fail in python 3.7 due to flask being unable to import escape
from jinja2. The fix is to pin jinja2 to a version prior to 3.1.0. More
details on this failure can be found in the github issue:
https://github.com/pallets/flask/issues/4494

CT-574

* Fix itsdangerous import that breaks py3.7 tests

Same issue as with jinja2. Pinning itsdangerous to the same version used
by py3.6 should fix the issue.

CT-574

* Force use of bash for run steps in GHA

GHA uses powershell by default on windows, and that is messing up shell
env evaluation.

CT-574

* Pin Werkzeug version in py3.7 as well

CT-574

* Don't run tests against EOL python 3.5

Fix the broken python 3.5 test suite by removing it. Python 3.5 has been
EOL for 2 years.

CT-574

* Add tox-gh-actions library for compatibility with GH actions

tox-gh-actions adds support for automatically detecting python version
and platform, as well as log grouping and annotating error messages.
https://github.com/ymyzk/tox-gh-actions

CT-574

* Customize packages installed on windows in python 3.10

We need to override the version of the pywin32 package installed on
windows in python 3.10. To accomplish this, I'm trying an experiment
using "factor-conditional" settings feature in tox. The package
tox-gh-actions configures factors based on job configuration, and then
we use the python version and platform factors to cause tox to install
the package version we need, where we need it.

Trying to do this in a manner that doesn't break the circleci tests.

CT-574

* Add tox-gh-actions to tox requires

This is necessary since we are using the `requires` setting in tox to
pin the virtualenv version. Without this change tox-gh-actions isn't
loaded.

CT-574

* Update platform factors to match matrix values

CT-574

* Enable deriving basepython from tox factors

CT-574

* Trying to fix unittests for python3.10 on windows

The wrong version of pywin32 is still being installed. Try to fix it by
using the latest pip version, and switching the factors around in deps.

CT-574

* Disable pinning of pip as it probably isn't needed anymore

CT-574

* Debugging experiment

turn up tox verbosity and try factor-conditionally installing a package
in linux and macos python 3.8.

CT-574

* Use full tox env name

Apparently testenvs aren't selected by factors like you can do with deps
or basepython, or anything else.

CT-574

* Unpin virtualenv

Virtualenv and pip were pinned two years ago, and are incredibly out of
date.

CT-574

* Remove dep factor test

Factor-conditional deps are now working.

CT-574

* Use different requirements files for windows and linux/darwin

CT-574

* Only run windows unittests on python 3.10

We only build the agent windows installer with python 3.10, so only test
against that python version so we can stop playing whack-a-mole with
windows unit tests.

CT-574

* Debugging & Cruft removal

- Remove verbosity flags from tox
- Remove commented out version pinning for pip and virtualenv. I have
  confirmed that tests work just fine with the latest virtualenv and pip
  in both GHA and CircleCI.

CT-574

* Coverage and split windows platform tests

Modify unittest commands to combine windows testenv configs and coverage
config all together. Coverage is now captured and windows platform tests
are run in a separate process (and only on windows environments).

The coverage data is saved to GHA artifacts, and then merged across all
tests and uploaded to codecov.

CT-574

* Include parent env vars

Fix issue where pytest output wasn't colorized.

CT-574

* Pin pycoverage version to fix mixed version warnings

CT-574

* Fixes & Requested Changes

- Set python version in coverage upload job.
- Only run unittests for macos on python 3.10 as requested.
- Run tests against python 2.7

CT-574

* Fix python 2.7 unittests job

Add python 2.7 to the list of envs so it is correctly handled.

CT-574

* Run publish-test-results only once

Having one comment per matrix combination is a bit much. Run the publish
test results in a separate job with the results of all tests.

CT-574

* Pin github action versions to SHA for security

CT-574

* Try caching Python dependencies to see if it speeds up tox run.

Co-authored-by: Tomaz Muraus <tomazm@sentinelone.com>