Fix unittest coverage on GHA
This fixes the following problems that were breaking coverage reporting
for the unittests that were run on GHA:
- The job that converts the coverage data to xml did not have access to
  the source code (which is necessary for it to work)
- The coverage combine job did not have the coveragerc config file
- The coveragerc config file does not have all necessary source paths
  listed, so combine fails
- The codecov config does not have a path fix for code from GHA runners

The first two issues were fixed by checking out the repo in the coverage
job. The third issue was fixed by adding all possible source paths to
.coveragerc config file. The last issue was fixed by adding GHA code
paths to the codecov config.

CT-573