Update pycoverage to 5.5
Pycoverage version 5 brings in a neat new feature that traces lines
covered back to the tests that cover them. This could be really useful
in debugging drops in coverage in the future. Version 5.5 is the last
version that supports python 3.5 and 2.7.

pytest-cov was also updated to support the updated coverage version.