Break up codespeed agent benchmark script over multiple steps
In an attempt to get debug output, and prevent `tail -f` from blocking
forever, break up the codespeed agent benchmarking script over multiple
steps.

The goal is to continue getting output from the
`start-agent-and-capture-metrics` script in the step that runs it,
despite the script being backgrounded so commands can be run after the
agent has been started.

Also ensure secrets are passed along to the reusable workflow.

CT-577