Update "query_object()" and "query_objects()" code to support fallback URL in case the first URL returns 404.
This is by no means a complete or robust replacement for a proper API
versioning and management strategy, but it allows us to relatively
easily and simply solve for the most common use case / problem when an
API path has been renamed or moved from v1beta to v1 (aka from beta to
stable such as /batch/v1beta/cronjobs endpoints in Kubernetes v1.25.0).

The code works by falling back to the next endpoint in the list in case
the first one returns 404. In case the first one returns success or some
other error, we return and propagate the error directly instead of
falling back to the next URL in the list.