[DSET-3928] [DSET-4094] Mid term approach for handling changed Kubernetes API URL endpoints in the Kubernetes monitors (#1143)
* Update "query_object()" and "query_objects()" code to support fallback
URL in case the first URL returns 404.

This is by no means a complete or robust replacement for a proper API
versioning and management strategy, but it allows us to relatively
easily and simply solve for the most common use case / problem when an
API path has been renamed or moved from v1beta to v1 (aka from beta to
stable such as /batch/v1beta/cronjobs endpoints in Kubernetes v1.25.0).

The code works by falling back to the next endpoint in the list in case
the first one returns 404. In case the first one returns success or some
other error, we return and propagate the error directly instead of
falling back to the next URL in the list.

* Introduce feature flag (k8s_fallback_urls_disable - enabled by default)
with which new fallback URLs functionality can be disabled for whatever
reason (e.g. large performance overhead or similar).

* Add changelog entry.

* Fix typo.

* Add tests for new fallback urls functionality.

* Fix typo.

* Fix feature flag off test case.

* Fix Python 2.x compatibility.

* Update Kubernetes end to end tests to also verify CronJob and Job events
are being correctly emitted by the Kubernetes Events monitor.

* Upgrade minikube.

* Also run checks against Kubernetes v1.25 and v1.26.

* Simplify the CI/CD conditional, skip check for event which is not always
present under all Kubernetes versions.

* Fix fallback urls handling and add test cases for scenario / URL without
the namespace since it utilizes a different code path.

* Also try running OM e2e tests against Kubernetes v1.25 and v1.26.

* Use latest patch versions of varions Kubernetes versions we run end to
end tests with.

* Add more assertions on the actual URLs which were used for the outgoing
HTTP requests.

* Add new config option which allows Kubernetes Events monitor to be
tested more easily and faster outside of Docker and Kubernetes.

Keep in mind that this feature is only meant to be used for local
development and testing and not to be used in production.

* Update affected tests.

* Add tests for kubernetes events monitor "skip_leader_election" config
option.

* Bump code-cov action to see if it helps with occasional temporary
failures we see.

* Update the script so it also works with coreutils on M1/M2 laptops which
use different homebrew path.

* Re-generate monitors docs.

* Test a workaround for a weird failure we started seeing recently.

* Disable the workaround which doesn't seem to be working.

* Try using different debian buster AMI for end to end tests.

* Remove testing changes which are not needed anymore.

* Fix accidental change and change log level back to warn as it originally
was.

* Update expected openssl version on debian 10.